+++ app/src/main/java/com/example/parkpal/domain/model/User.kt
@@ -1,7 +1,7 @@
 package com.example.parkpal.domain.model
 
 data class User(
-    val userId: Int = 0,
+    val userId: Long = 0,
     val name: String,
     val email: String,
     val password: String?,
